重新定了设计,简单的修改了一下布局

+++ EnergyCycles/Tool/Help/CacheManager.h
@@ -15,7 +15,7 @@
  */
 
 
-- (long long)fileSizeAtPath:(NSString *)path;
+- (CGFloat)fileSizeAtPath:(NSString *)path;
 
 - (CGFloat)folderSizeAtPath:(NSString *)path;
